requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
# Note: Using psycopg3 instead of asyncpg for better Windows compatibility
psycopg[binary]==3.3.2

# HTTP Client
httpx==0.26.0

# Configuration
python-dotenv==1.0.0

# Templates
jinja2==3.1.3
aiofiles==23.2.1

# Utilities
python-multipart==0.0.6
